making positional fields 'doubles' making r,cp,g constants/parameters testing ecDet

+++ src/common/particleML.f90
@@ -13,9 +13,9 @@ module particleML
 !           pdata(8,n) - precipitation intensity (mm/hour)
 !           pdata(9,n) - radioactive content (Bq)
 !           pdata(10,n) - gravity in m/s (fixed or computed)
-       REAL    :: x     !- x position in grid
-       REAL    :: y     !- y position in grid
-       REAL    :: z     !- sigma/eta position (vertical)
+       REAL*8    :: x     !- x position in grid
+       REAL*8    :: y     !- y position in grid
+       REAL*8    :: z     !- sigma/eta position (vertical)
        REAL    :: tbl   !- sigma/eta at top of boundary layer
        REAL    :: rad   !- radioactive content (Bq)
        REAL    :: hbl
@@ -28,8 +28,8 @@ module particleML
     TYPE extraParticle
        REAL    :: u ! u-speed
        REAL    :: v ! v-speed
-       REAL    :: rmx ! map ratio in x direction
-       REAL    :: rmy ! map ration in y direction
+       REAL*8  :: rmx ! map ratio in x direction
+       REAL*8  :: rmy ! map ration in y direction
        REAL    :: prc ! precipition intensity (mm/hour)
     END TYPE extraParticle
